+++ AzeriteUI/back-end/libraries/frame.lua
@@ -1,4 +1,4 @@
-local LibFrame = Wheel:Set("LibFrame", 68)
+local LibFrame = Wheel:Set("LibFrame", 69)
 if (not LibFrame) then	
 	return
 end
@@ -61,6 +61,10 @@ local KEYWORD_DEFAULT = "UICenter"
 -- Create the new frame parent 
 if (not LibFrame.frameParent) then
 	LibFrame.frameParent = LibFrame.frameParent or CreateFrame("Frame", nil, UIParent, "SecureHandlerAttributeTemplate")
+	RegisterAttributeDriver(LibFrame.frameParent, "state-visibility", "[combat]show;show")
+else
+	UnregisterAttributeDriver(LibFrame.frameParent, "state-visibility")
+	RegisterAttributeDriver(LibFrame.frameParent, "state-visibility", "[combat]show;show")
 end 
 
 -- Create the cache frame
@@ -78,12 +82,6 @@ else
 	LibFrame.frame:SetParent(LibFrame.frameParent) 
 	UnregisterAttributeDriver(LibFrame.frame, "state-visibility")
 end 
-
--- Hide the master visibility frame if we haven't yet reached login. 
--- This might improve addon loading time. 
---if (not IsLoggedIn()) and (not InCombatLockdown()) then 
---	LibFrame.frameParent:Hide()
---end
 
 -- Return a value rounded to the nearest integer.
 local math_round = function(value)
@@ -440,29 +438,12 @@ LibFrame.OnEvent = function(self, event, ...)
 	self:UpdateDisplaySize()
 end
 
-LibFrame.OnReload = function(self, event, ...)
-	if (event == "PLAYER_LEAVING_WORLD") then 
-		if (not InCombatLockdown()) then 
-			VisibilityFrame:Hide()
-		end
-	elseif (event == "PLAYER_ENTERING_WORLD") then 
-		if (not VisibilityFrame:IsShown()) then 
-			VisibilityFrame:Show()
-		end
-	end
-end
-
 LibFrame.Enable = function(self)
 
 	-- Get rid of old events from previous handlers, 
 	-- if this library for some reason was overwritten 
 	-- by a more recent version from a load on demand addon. 
 	self:UnregisterAllEvents()
-
-	-- Hide the visibility frame when reloading
-	-- The idea is to just stop all running OnUpdate handlers
-	--self:RegisterEvent("PLAYER_ENTERING_WORLD", "OnReload")
-	--self:RegisterEvent("PLAYER_LEAVING_WORLD", "OnReload")
 
 	-- New system only needs to capture changes and events
 	-- affecting display size or the cinematic frame visibility.